全状態の統合、Injectへの変更
