+++ Assets/App/Scripts/Battle/BattleLifetimeScope.cs
@@ -1,8 +1,8 @@
 using App.Battle.Datastores;
 using App.Battle.UseCases;
+using App.Battle.Presenters;
 using App.Battle.Interfaces.UseCases;
 using App.Battle.Interfaces.Datastores;
-
 using UnityEngine;
 using VContainer;
 using VContainer.Unity;
@@ -13,5 +13,7 @@ public class BattleLifetimeScope : LifetimeScope
     {
         builder.Register<BattleEnemyDatastore>(Lifetime.Singleton).AsImplementedInterfaces();
         builder.Register<BattleEnemyUseCase>(Lifetime.Singleton).AsImplementedInterfaces();
+        builder.Register<BattleStateMachineUseCase>(Lifetime.Singleton).AsImplementedInterfaces();
+        builder.Register<BattleStateMachinePresenter>(Lifetime.Singleton).AsImplementedInterfaces();
     }
 }
