+++ Assets/App/Scripts/Battle/Interfaces/Presenters/IBattleStateMachinePresenter.cs
@@ -1,0 +1,10 @@
+namespace App.Battle.Interfaces.Presenters
+{
+    public interface IBattleStateMachinePresenter
+    {
+        public void IncreaseIndex();
+        public void DecreaseIndex();
+        public void Execute();
+        public void Cancel();
+    }
+}
