+++ Assets/App/Scripts/Battle/Interfaces/Presenters/IBattleStateMachinePresenter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ed309757fbcac814783dc6d340e7364c
+guid: 1eb4311cd0f1c0c41a184b4093089299
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
