+++ Assets/App/Scripts/Debug/Battle/BattleDebugLifetimeScope.cs
@@ -1,6 +1,5 @@
 using App.Debug.Battle.Presenters;
 using App.Debug.Battle.UseCases;
-using App.Battle.UseCases;
 using UnityEngine;
 using VContainer;
 using VContainer.Unity;
@@ -10,34 +9,14 @@ namespace App.Debug.Battle
     public class BattleDebugLifetimeScope : LifetimeScope
     {
         [SerializeField] private BattleDebugEnemyPresenter _BattleDebugEnemyPresenter;
-        [SerializeField] private BattleDebugExecuteStatePresenter _BattleDebugExecuteStatePresenter;
-        [SerializeField] private BattleDebugCancelStatePresenter _BattleDebugCancelStatePresenter;
-        [SerializeField] private BattleDebugIncreaseStateIndexPresenter _BattleDebugIncreaseStateIndexPresenter;
-        [SerializeField] private BattleDebugDecreaseStateIndexPresenter _BattleDebugDecreaseStateIndexPresenter;
+        [SerializeField] private BattleDebugStatePresenter _BattleDebugStatePresenter;
 
         protected override void Configure(IContainerBuilder builder)
         {
             builder.RegisterComponent(_BattleDebugEnemyPresenter).AsImplementedInterfaces();
-
             builder.RegisterEntryPoint<BattleDebugEnemyUseCase>();
-
-            builder.RegisterComponent(_BattleDebugExecuteStatePresenter).AsImplementedInterfaces();
-
-            builder.RegisterEntryPoint<BattleDebugExecuteStateUseCase>();
-
-            builder.RegisterComponent(_BattleDebugCancelStatePresenter).AsImplementedInterfaces();
-
-            builder.RegisterEntryPoint<BattleDebugCancelStateUseCase>();
-
-            builder.RegisterComponent(_BattleDebugIncreaseStateIndexPresenter).AsImplementedInterfaces();
-
-            builder.RegisterEntryPoint<BattleDebugIncreaseStateIndexUseCase>();
-
-            builder.RegisterComponent(_BattleDebugDecreaseStateIndexPresenter).AsImplementedInterfaces();
-
-            builder.RegisterEntryPoint<BattleDebugDecreaseStateIndexUseCase>();
-
-            builder.Register<BattleStateMachineUseCase>(Lifetime.Singleton).AsImplementedInterfaces();
+            builder.RegisterComponent(_BattleDebugStatePresenter).AsImplementedInterfaces();
+            builder.RegisterEntryPoint<BattleDebugStateUseCase>();
         }
     }
 }
