--- Assets/App/Scripts/Debug/Battle/Interfaces/Presenters/IBattleDebugCancelStatePresenter.cs
@@ -1,10 +1,0 @@
-using System;
-using UniRx;
-
-namespace App.Debug.Battle.Interfaces.Presenters
-{
-    public interface IBattleDebugCancelStatePresenter
-    {
-        IObservable<Unit> OnCancelState { get; }
-    }
-}
