--- Assets/App/Scripts/Debug/Battle/Interfaces/Presenters/IBattleDebugCancelStatePresenter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f84370d3c7e36eb40a67ef3d56f1665f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
