--- Assets/App/Scripts/Debug/Battle/Interfaces/Presenters/IBattleDebugDecreaseStateIndexPresenter.cs
@@ -1,10 +1,0 @@
-using System;
-using UniRx;
-
-namespace App.Debug.Battle.Interfaces.Presenters
-{
-    public interface IBattleDebugDecreaseStateIndexPresenter
-    {
-        IObservable<Unit> OnDecreaseStateIndex { get; }
-    }
-}
