--- Assets/App/Scripts/Debug/Battle/Interfaces/Presenters/IBattleDebugDecreaseStateIndexPresenter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fe2124bbf66c99443ba9ccd95e05977a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
