--- Assets/App/Scripts/Debug/Battle/Interfaces/Presenters/IBattleDebugExecuteStatePresenter.cs
@@ -1,10 +1,0 @@
-using System;
-using UniRx;
-
-namespace App.Debug.Battle.Interfaces.Presenters
-{
-    public interface IBattleDebugExecuteStatePresenter
-    {
-        IObservable<Unit> OnExecuteState { get; }
-    }
-}
