--- Assets/App/Scripts/Debug/Battle/Interfaces/Presenters/IBattleDebugExecuteStatePresenter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 20b965044fb6a604f9660205b58385f3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
