--- Assets/App/Scripts/Debug/Battle/Interfaces/Presenters/IBattleDebugStateIncreaseStateIndexPresenter.cs
@@ -1,10 +1,0 @@
-using System;
-using UniRx;
-
-namespace App.Debug.Battle.Interfaces.Presenters
-{
-    public interface IBattleDebugIncreaseStateIndexPresenter
-    {
-        IObservable<Unit> OnIncreaseStateIndex { get; }
-    }
-}
