--- Assets/App/Scripts/Debug/Battle/Interfaces/Presenters/IBattleDebugStateIncreaseStateIndexPresenter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8b4a2728ad43d7a47a7a245ecf992dfc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
