+++ Assets/App/Scripts/Debug/Battle/Interfaces/Presenters/IBattleDebugStatePresenter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0b81a831a72bb454c952ff010688b5b0
+guid: 40269e902a7d2af44bbd369cea3468cd
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
