--- Assets/App/Scripts/Debug/Battle/Presenters/BattleDebugCancelStatePresenter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 23de40b3f37236c45857e30018028ebf
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
