--- Assets/App/Scripts/Debug/Battle/Presenters/BattleDebugDecreaseStateIndexPresenter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9a7bf9fad77930e4c87150da8a2be263
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
