--- Assets/App/Scripts/Debug/Battle/Presenters/BattleDebugExecuteStatePresenter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 96915dce1f4e2d6429445aa22fd66fa1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
