+++ Assets/App/Scripts/Debug/Battle/Presenters/BattleDebugStatePresenter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c8875b5a73aab2a4d835a57e9e9c45b6
+guid: e34d606a49d146b4094830e672c1588f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
