--- Assets/App/Scripts/Debug/Battle/UseCases/BattleDebugCancelStateUseCase.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c4df702301d78064f86fa2c526fbbde7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
